COMCOM-014: Completed the front-end portion of the ticket in the-observatory
The changes include an added button and linked method in the order-card component, an added modal and some methods in the order component, added CSS to go with the aforementioned changes, a new Order Service API call method, and a new Cypress test case. Other changes made include an addition of a missing log message to the OrderService in meteor-shower, and the deletion of an unused module import the comic-book-card component in sage-mountain.