COMCOM-015 -  Contact Us (#15)
* COMCOM-015: created the back-end portion of this ticket, shooting-star

I also  added a new section to the README covering initial setup for that service. In addition, there were  minor changes to the TokenManager files in both meteor-shower and sage-cave.

* COMCOM-015: Finished the front-end part of the ticket in the-observatory

This includes minor style changes for files, as well as a new endpoint, a new Cypress test suite, three new unit test files and a new component folder. There was also a minor change to the validator class in shooting-star.

* PR CHANGES: formatting change